Fixes compatibility with Laravel 5.6
This change uses a Symfony Dependency Injection Reference to the Laravel
app, which is registered in
`Laracasts\Behat\ServiceContainer\BehatExtension@loadLaravel`, instead
of injecting the app object directly. Doing so allows `serialize` calls,
present in symfony/dependency-injection 4.x, to serialize the
LaravelArgumentResolver without throwing an exception.

Please note, due to to a similar issue in Behat, Laravel 5.6 will only
work with this package when the Behat install has the changes in
flaxandteal/Behat:bug/1108-serialization-of-closure-is-not-allowed